Don't destroy servers in SG before deleting the SG from a scheduler.
When a server group is deleted from a scheduler, servers in the group
lose their association with the scheduler as well. 'sched_data' of
each server is NULLed to mark that a server is no longer linked with
a scheduler. Technically, that makes it possible to add the group to
a different scheduler.

When a server group is destroyed, delete it from a scheduler first,
and destroy servers in the group afterwards.